start splitting datepicker apart
